+++ media/gpu/h264_rate_controller.cc
@@ -17,7 +17,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 namespace media {
 namespace {
 // Base temporal layer index.
-constexpr int kBaseLayerIndex = 0;
+constexpr size_t kBaseLayerIndex = 0;
 
 // Delta QP between layers in Fixed Delta QP mode. It is arbitrary chosen value.
 constexpr int kFixedLayerDeltaQP = 4;
